float edit

+++ Week1/Lab1/Primitives.java
@@ -5,7 +5,7 @@ public class Primitives {
         short minShort = -32768;
         int maxInt = 2147483647;
         long javaFlex = 3000000000L;
-        float pie = (float) 3.14159;
+        float pie = 3.14159F;
         double doubleYum = 6.28318;
